+++ python/megane/__init__.py
@@ -4,6 +4,7 @@ from megane.parsers.cif import load_cif
 from megane.parsers.lammps_data import load_lammps_data
 from megane.parsers.lammpstrj import load_lammpstrj_structure
 from megane.parsers.pdb import load_pdb
+from megane.parsers.phonon import load_phonon
 from megane.parsers.traj import load_traj
 from megane.parsers.vasp import load_vasp
 from megane.parsers.xtc import load_trajectory
@@ -56,6 +57,7 @@ __all__ = [
     "load_lammps_data",
     "load_lammpstrj_structure",
     "load_pdb",
+    "load_phonon",
     "load_traj",
     "load_trajectory",
     "load_vasp",